student/discord_bot/pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
]
